debugging

+++ libft/ft_power_db.c
@@ -6,15 +6,15 @@
 /*   By: vhazelnu <vhazelnu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/07 15:30:32 by vhazelnu          #+#    #+#             */
-/*   Updated: 2019/07/07 15:46:15 by vhazelnu         ###   ########.fr       */
+/*   Updated: 2019/07/08 10:19:59 by vhazelnu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_power_db(double n, int p)
+long double		ft_power_db(long double n, int p)
 {
-	double	result;
+	long double		result;
 
 	result = 1;
 	if (p < 0)
